refactor: 폴더 및 파일 명, 폴더 구조 변경
