+++ src/main/java/gift/entity/Product.java
@@ -1,4 +1,4 @@
-package gift.model;
+package gift.entity;
 
 public class Product {
