+++ src/test/java/gift/ProductTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import gift.dto.ProductDTO;
-import gift.model.Product;
+import gift.domain.product.ProductDTO;
+import gift.entity.Product;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
